+++ src/ts/Ads/Ads.ts
@@ -39,7 +39,7 @@ import { ProgrammaticTrackingService } from 'Ads/Utilities/ProgrammaticTrackingS
 import { SdkStats } from 'Ads/Utilities/SdkStats';
 import { SessionDiagnostics } from 'Ads/Utilities/SessionDiagnostics';
 import { InterstitialWebPlayerContainer } from 'Ads/Utilities/WebPlayer/InterstitialWebPlayerContainer';
-import { NewVideoOverlay } from 'Ads/Views/NewVideoOverlay';
+import { VideoOverlay } from 'Ads/Views/VideoOverlay';
 import { Banners } from 'Banners/Banners';
 import { AuctionRequest } from 'Banners/Utilities/AuctionRequest';
 import { FinishState } from 'Core/Constants/FinishState';
@@ -514,7 +514,7 @@ export class Ads implements IAds {
         }
 
         if(TestEnvironment.get('autoSkip')) {
-            NewVideoOverlay.setAutoSkip(TestEnvironment.get('autoSkip'));
+            VideoOverlay.setAutoSkip(TestEnvironment.get('autoSkip'));
         }
 
         if(TestEnvironment.get('autoClose')) {